requirements.txt
google-genai>=0.3.0
google-auth>=2.23.0
google-cloud-storage>=2.10.0
pyyaml>=6.0
tqdm>=4.65.0
